Removed the round overloaded function.
In Linux, there is already a round(double) function, but it returns
another double instead of an integer, so it created a conflict.

I found that the round is only used a couple of times to convert a
rounded double to an integer by adding 0.5 and then truncating the value.
Instead of using this function, I just added the 0.5 myself on the
couple of places that needs it.